requirements.txt
langchain
python-dotenv
ipykernel
langchain_community
langchain-openai
langchain-chroma>=0.1.2
docx2txt
sentence_transformers
langchain_huggingface
faiss-cpu